ovs_appctl_compgen: Solved a very big issue.
With compgen, user needs to press [tab] twice to get the completion
of a new argument.  The first press of [tab] will not show anything.
However, since the current script contains several 'printf' to stdout,
the first press of [tab] will cause the print of those helping
messages.  This can be very confusing!

To avoid this, I found that the compgen fills in different values in
the $COMP_TYPE for each [tab] pressed, 9 and 63, respectively.
Although I need to investigate further about the meaning of COMP_TYPE
code, $COMP_TYPE can be used as indicator to suppress the 'printf'
output at first [tab].

So, this commit takes advantage of $COMP_TYPE and makes the compgen
function only print to stdout at second [tab] execution.

NOTE:
- It is hacky, but I'm super happy!

Signed-off-by: Yeqiao Wang <ee07b291@gmail.com>