ovs_appctl_compgen: Refactor the design.
After a while of using, the current design of parsing the
manpage takes too long to finish the completion of long
sub-commands.

Therefore, decision is made that a 2.0 branch be created.
The new design will directly call the ovs-appctl cmd to
get the sub-commands and their formats.  Also, the parsing
code will no longer calculate all combinations, instead,
it will take in one input argument at a time and gradually
look for next possible completions.

Note: the code is not finished.  A working commit will be
marked soon.

Signed-off-by: Yeqiao Wang <ee07b291@gmail.com>